Rename onvif event module to event_manager.
This conflicts with creating the platform event, which is now needed.
