+++ homeassistant/components/onvif/event_manager.py
