+++ tests/components/onvif/test_event_manager.py
